Moves Team types to types.ts
