+++ src/components/Contributors/Contributors.tsx
@@ -5,8 +5,10 @@ import Link from '@material-ui/core/Link';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 
+import { Team, TeamMember } from 'types';
+
 interface Props {
-    team: Prytaneum.Team;
+    team: Team;
 }
 
 const useStyles = makeStyles((theme: Theme) =>
@@ -20,7 +22,7 @@ const useStyles = makeStyles((theme: Theme) =>
 export default function Contributors({ team }: Props) {
     const classes = useStyles();
 
-    const [subTeam, setSubTeam] = React.useState<Prytaneum.TeamMember[]>([]);
+    const [subTeam, setSubTeam] = React.useState<TeamMember[]>([]);
 
     React.useEffect(() => {
         setSubTeam(team.members.slice(0, 10));
