+++ src/types.ts
@@ -1,0 +1,20 @@
+interface Reference {
+    link: string;
+    name: string;
+    icon: JSX.Element;
+}
+
+export interface TeamMember {
+    picturePath?: string;
+    fullName: string;
+    subtitle: string;
+    description: string;
+    startDate: Date;
+    endDate: Date;
+    references?: Reference[];
+}
+
+export interface Team {
+    name: string;
+    members: TeamMember[];
+}
